refactor: `Input` export
