+++ client/src/lib/components/ui/index.ts
@@ -1,6 +1,6 @@
 export { default as Button } from "./Button/Button.svelte";
 export { default as Card } from "./Card/Card.svelte";
-export { default as Input } from "./TextInput/Input.svelte";
+export { default as Input } from "./Input/Input.svelte";
 export { default as Modal } from "./Modal/Modal.svelte";
 export { default as Avatar } from "./Avatar/Avatar.svelte";
 export * from "./Table";
